requirements.txt
torch
torchvision
transformers
opencv-python
pillow
numpy
matplotlib
tqdm
gradio
ultralytics
segment-anything
